feat(be): separate response into package
