+++ apps/iris/src/router/response/judge.go
@@ -1,7 +1,6 @@
-package router
+package response
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"strconv"
@@ -9,14 +8,14 @@ import (
 	"github.com/skkuding/codedang/apps/iris/src/handler"
 )
 
-type Response struct {
+type JudgeResponse struct {
 	SubmissionId    int                `json:"submissionId"`
 	JudgeResultCode handler.ResultCode `json:"resultCode"`
 	JudgeResult     json.RawMessage    `json:"judgeResult"`
 	Error           string             `json:"error"`
 }
 
-func NewResponse(id string, data json.RawMessage, err error) *Response {
+func NewJudgeResponse(id string, data json.RawMessage, err error) *JudgeResponse {
 	resultCode := handler.ACCEPTED
 	errMessage := ""
 
@@ -30,7 +29,7 @@ func NewResponse(id string, data json.RawMessage, err error) *Response {
 	}
 
 	_id, _ := strconv.Atoi(id)
-	return &Response{
+	return &JudgeResponse{
 		SubmissionId:    _id,
 		JudgeResultCode: resultCode,
 		JudgeResult:     data,
@@ -38,16 +37,7 @@ func NewResponse(id string, data json.RawMessage, err error) *Response {
 	}
 }
 
-func JSONMarshal(t interface{}) ([]byte, error) {
-	// source: https://stackoverflow.com/questions/28595664/how-to-stop-json-marshal-from-escaping-and
-	buffer := &bytes.Buffer{}
-	encoder := json.NewEncoder(buffer)
-	encoder.SetEscapeHTML(false)
-	err := encoder.Encode(t)
-	return buffer.Bytes(), err
-}
-
-func (r *Response) Marshal() []byte {
+func (r *JudgeResponse) Marshal() []byte {
 
 	if res, err := JSONMarshal(r); err != nil {
 		// Error on marshaling router response means that
